Frontend testing fixes

+++ src/main/java/model/frontend/FrontEndSearchRequest.java
@@ -2,14 +2,14 @@ package model.frontend;
 
 public class FrontEndSearchRequest {
     private String searchQuery;
-    private int maxResults = Integer.MAX_VALUE;
+    private long maxResults = Long.MAX_VALUE;
     private double minScore = 0.0;
 
     public String getSearchQuery() {
         return searchQuery;
     }
 
-    public int getMaxResults() {
+    public long getMaxResults() {
         return maxResults;
     }
